Do not use yq before install and protect directory JPHUB to remove from bas conf

+++ install/install-system.sh
@@ -6,35 +6,35 @@ date
 
 export WODTYPE=$1
 if [ -z "$WODTYPE" ]; then
-	echo "Syntax: install-system.sh api-db|backend|frontend|appliance"
-	exit -1
+    echo "Syntax: install-system.sh api-db|backend|frontend|appliance"
+    exit -1
 fi
 
 launch_with_pm2() {
-	DIR=$1
-	shift
-	APP=$1
-	shift
-	echo "Install pm2"
-	npm install pm2@latest
-	export PATH=$PATH:"$DIR/node_modules/pm2/bin"
-	cat >> $HOME/.bash_profile << EOF
+    DIR=$1
+    shift
+    APP=$1
+    shift
+    echo "Install pm2"
+    npm install pm2@latest
+    export PATH=$PATH:"$DIR/node_modules/pm2/bin"
+    cat >> $HOME/.bash_profile << EOF
 export PATH=$PATH:"$DIR/node_modules/pm2/bin"
 EOF
-	# Allow error to occur
-	set +e
-	pm2 show $APP 2>&1 > /dev/null
-	if [ $? -eq 0 ]; then
-		echo "Stop a previous server for $APP"
-		pm2 del $APP
-	fi
-	set -e
-	echo "Start the $APP server"
-	pm2 start --name=$APP npm -- start
+    # Allow error to occur
+    set +e
+    pm2 show $APP 2>&1 > /dev/null
+    if [ $? -eq 0 ]; then
+        echo "Stop a previous server for $APP"
+        pm2 del $APP
+    fi
+    set -e
+    echo "Start the $APP server"
+    pm2 start --name=$APP npm -- start
 }
 
 if [ ! -f $HOME/.gitconfig ]; then
-	cat > $HOME/.gitconfig << EOF
+    cat > $HOME/.gitconfig << EOF
 # This is Git's per-user configuration file.
 [user]
 # Please adapt and uncomment the following lines:
@@ -99,8 +99,8 @@ export WODPRIVNOBO=$WODPRIVDIR/notebooks
 WODPRIVINV=""
 # Manages private inventory if any
 if [ -f $WODPRIVDIR/ansible/inventory ]; then
-	WODPRIVINV="-i $WODPRIVDIR/ansible/inventory"
-	export WODPRIVINV
+    WODPRIVINV="-i $WODPRIVDIR/ansible/inventory"
+    export WODPRIVINV
 fi
 
 # AIP-DB PART
@@ -111,7 +111,7 @@ export WODFEDIR=$PWODBEDIR/wod-frontend
 export WODNOBO=$PWODBEDIR/wod-notebooks
 EOF
 if [ $WODTYPE = "backend" ]; then
-	cat >> $SCRIPTDIR/wod.sh << 'EOF'
+    cat >> $SCRIPTDIR/wod.sh << 'EOF'
 
 # These dirs are also fixed by default and can be changed as needed
 export STUDDIR=/student
@@ -126,9 +126,21 @@ cd $ANSIBLEDIR
 PBKDIR=$WODGROUP
 
 if [ $WODTYPE = "backend" ]; then
-	export JPHUB=`cat "$ANSIBLEDIR/group_vars/$PBKDIR" | yq '.JPHUB' | sed 's/"//g'`
-	# In case of update remove first old jupyterhub version
-	sudo rm -rf $JPHUB
+    JPHUB=`cat "$ANSIBLEDIR/group_vars/$PBKDIR" | grep -E '^JPHUB:' | cut -d: -f2`
+    # In case of update remove first old jupyterhub version
+    if [ _"$JPHUB" = _"" ]; then
+        echo "Directory for jupyterhub is empty"
+        exit -1
+    fi
+    if [ _"$JPHUB" = _"/" ]; then
+        echo "Directory for jupyterhub is /"
+        exit -1
+    fi
+    if [ _"$JPHUB" = _"$HOME" ]; then
+        echo "Directory for jupyterhub is $HOME"
+        exit -1
+    fi
+    sudo rm -rf $JPHUB
 fi
 
 
@@ -149,11 +161,11 @@ then
     exit -1
 fi
 if [ $WODDISTRIB = "centos-7" ] || [ $WODDISTRIB = "ubuntu-20.04" ] || [ $WODDISTRIB = "ubuntu-22.04" ]; then
-	# Older distributions require an older version of the collection to work.
-	# See https://github.com/ansible-collections/community.general
-	ansible-galaxy collection install --force-with-deps community.general:4.8.5
+    # Older distributions require an older version of the collection to work.
+    # See https://github.com/ansible-collections/community.general
+    ansible-galaxy collection install --force-with-deps community.general:4.8.5
 else
-	ansible-galaxy collection install community.general
+    ansible-galaxy collection install community.general
 fi
 ansible-galaxy collection install ansible.posix
 
@@ -161,16 +173,16 @@ ansible-galaxy collection install ansible.posix
 SCRIPTREL=`echo $SCRIPT | perl -p -e "s|$WODBEDIR||"`
 if [ -x $WODPRIVDIR/$SCRIPTREL ];
 then
-	echo "Executing additional private script $WODPRIVDIR/$SCRIPTREL"
-	$WODPRIVDIR/$SCRIPTREL
+    echo "Executing additional private script $WODPRIVDIR/$SCRIPTREL"
+    $WODPRIVDIR/$SCRIPTREL
 fi
 
 ANSPRIVOPT=""
 if [ -f "$ANSIBLEPRIVDIR/group_vars/all.yml" ]; then
-	ANSPRIVOPT="$ANSPRIVOPT -e @$ANSIBLEPRIVDIR/group_vars/all.yml"
+    ANSPRIVOPT="$ANSPRIVOPT -e @$ANSIBLEPRIVDIR/group_vars/all.yml"
 fi
 if [ -f "$ANSIBLEPRIVDIR/group_vars/$PBKDIR" ]; then
-	ANSPRIVOPT="$ANSPRIVOPT -e @$ANSIBLEPRIVDIR/group_vars/$PBKDIR"
+    ANSPRIVOPT="$ANSPRIVOPT -e @$ANSIBLEPRIVDIR/group_vars/$PBKDIR"
 fi
 # For future wod.sh usage by other scripts
 cat >> $SCRIPTDIR/wod.sh << EOF
@@ -179,25 +191,25 @@ EOF
 export ANSPRIVOPT
 
 if [ $WODTYPE = "backend" ]; then
-	ANSPLAYOPT="$ANSPLAYOPT -e LDAPSETUP=0 -e APPMIN=0 -e APPMAX=0"
+    ANSPLAYOPT="$ANSPLAYOPT -e LDAPSETUP=0 -e APPMIN=0 -e APPMAX=0"
 elif [ $WODTYPE = "api-db" ] || [ $WODTYPE = "frontend" ]; then
-	ANSPLAYOPT="$ANSPLAYOPT -e LDAPSETUP=0"
+    ANSPLAYOPT="$ANSPLAYOPT -e LDAPSETUP=0"
 fi
 
 # Inventory based on the installed system
 if [ $WODTYPE = "backend" ]; then
-	# In this case WODBEFQDN represents a single system
-	cat > $ANSIBLEDIR/inventory << EOF
+    # In this case WODBEFQDN represents a single system
+    cat > $ANSIBLEDIR/inventory << EOF
 [$WODGROUP]
 $WODBEFQDN ansible_connection=local
 EOF
 elif [ $WODTYPE = "api-db" ]; then
-	cat > $ANSIBLEDIR/inventory << EOF
+    cat > $ANSIBLEDIR/inventory << EOF
 [$WODGROUP]
 $WODAPIDBFQDN ansible_connection=local
 EOF
 elif [ $WODTYPE = "frontend" ]; then
-	cat > $ANSIBLEDIR/inventory << EOF
+    cat > $ANSIBLEDIR/inventory << EOF
 [$WODGROUP]
 $WODFEFQDN ansible_connection=local
 EOF
@@ -207,8 +219,8 @@ fi
 export USERMAX=`ansible-inventory -i $ANSIBLEDIR/inventory $WODPRIVINV --host $PBKDIR --playbook-dir $ANSIBLEDIR --playbook-dir $ANSIBLEPRIVDIR | jq ".USERMAX"`
 
 if [ $WODTYPE != "appliance" ]; then
-	# Setup this using the group for WoD
-	cat > $ANSIBLEDIR/group_vars/$WODGROUP << EOF
+    # Setup this using the group for WoD
+    cat > $ANSIBLEDIR/group_vars/$WODGROUP << EOF
 PBKDIR: $WODGROUP
 # 
 # Installation specific values
@@ -225,17 +237,17 @@ WODFEPORT: $WODFEPORT
 WODAPIDBPORT: $WODAPIDBPORT
 WODPOSTPORT: $WODPOSTPORT
 EOF
-	cat $ANSIBLEDIR/group_vars/wod-system >> $ANSIBLEDIR/group_vars/$WODGROUP
+    cat $ANSIBLEDIR/group_vars/wod-system >> $ANSIBLEDIR/group_vars/$WODGROUP
 
-	if [ -f $ANSIBLEDIR/group_vars/wod-$WODTYPE ]; then
-		cat $ANSIBLEDIR/group_vars/wod-$WODTYPE >> $ANSIBLEDIR/group_vars/$WODGROUP
-	fi
+    if [ -f $ANSIBLEDIR/group_vars/wod-$WODTYPE ]; then
+        cat $ANSIBLEDIR/group_vars/wod-$WODTYPE >> $ANSIBLEDIR/group_vars/$WODGROUP
+    fi
 fi
 
 if [ $WODTYPE = "backend" ]; then
-	# Compute WODBASESTDID based on the number that this backend server is multiplied by the number of users wanted
-	WODBASESTDID=$(($USERMAX*$WODBENBR))
-	cat >> $ANSIBLEDIR/group_vars/$WODGROUP << EOF
+    # Compute WODBASESTDID based on the number that this backend server is multiplied by the number of users wanted
+    WODBASESTDID=$(($USERMAX*$WODBENBR))
+    cat >> $ANSIBLEDIR/group_vars/$WODGROUP << EOF
 #
 # WODBASESTDID is the offset used to create users in the DB. It is required that each backend has a different non overlapping value.
 # Overlap is defined by adding USERMAX (from all.yml)
@@ -253,29 +265,29 @@ EOF
 fi
 
 if [ $WODTYPE != "appliance" ]; then
-	# Automatic Installation script for the system 
-	ansible-playbook -i inventory $WODPRIVINV --limit $PBKDIR $ANSPLAYOPT $ANSPRIVOPT install_$WODTYPE.yml
-	if [ $? -ne 0 ]; then
-		echo "Install had errors exiting before launching startup"
-		exit -1
-	fi
-	if [ $WODTYPE = "api-db" ]; then
-		get_wodapidb_userpwd
-		export PGPASSWORD="TrèsCompliqué!!##123"
-	fi
+    # Automatic Installation script for the system 
+    ansible-playbook -i inventory $WODPRIVINV --limit $PBKDIR $ANSPLAYOPT $ANSPRIVOPT install_$WODTYPE.yml
+    if [ $? -ne 0 ]; then
+        echo "Install had errors exiting before launching startup"
+        exit -1
+    fi
+    if [ $WODTYPE = "api-db" ]; then
+        get_wodapidb_userpwd
+        export PGPASSWORD="TrèsCompliqué!!##123"
+    fi
 fi
 
 if [ $WODTYPE = "api-db" ]; then
-	# We can now generate the seeders files 
-	# for the api-db server using the backend content installed as well
+    # We can now generate the seeders files 
+    # for the api-db server using the backend content installed as well
 
-	$INSTALLDIR/build-seeders.sh
+    $INSTALLDIR/build-seeders.sh
 
-	cd $WODAPIDBDIR
-	echo "Launching npm install..."
-	npm install
+    cd $WODAPIDBDIR
+    echo "Launching npm install..."
+    npm install
 
-	cat > .env << EOF
+    cat > .env << EOF
 FROM_EMAIL_ADDRESS="$WODSENDER"
 DB_PW=$PGPASSWORD
 DURATION=4
@@ -294,48 +306,49 @@ SENDGRID_API_KEY="None"
 WODUID=`id -u`
 WODGID=`id -g`
 EOF
-	echo "Launching docker PostgreSQL stack"
-	# Start the PostgreSQL DB stack
-	PGSQLDIR=`cat $WODAPIDBDIR/docker-compose.yml | yq '.services.db.environment.PGDATA' | sed 's/"//g'`
-	# We need to relog with sudo as $WODUSER so it's really in the docker group
-	# and be able to communicate with docker
-	# and we need to stop it before to be idempotent
-	# and we need to remove the data directory not done by the compose down
-	sudo su - $WODUSER -c "cd $WODAPIDBDIR ; docker compose down"
-	# That dir is owned by lxd, so needs root to remove
-	sudo su - -c "rm -rf $PGSQLDIR"
-	sudo su - $WODUSER -c "cd $WODAPIDBDIR ; docker compose config ; docker compose up -d"
-	POSTGRES_DB=`cat $WODAPIDBDIR/docker-compose.yml | yq '.services.db.environment.POSTGRES_DB' | sed 's/"//g'`
-	echo "Reset DB data"
-	npm run reset-data
-	echo "Setup user $WODAPIDBUSER"
-	psql --dbname=$POSTGRES_DB --username=postgres --host=localhost -c 'CREATE EXTENSION IF NOT EXISTS pgcrypto;'
-	psql --dbname=$POSTGRES_DB --username=postgres --host=localhost -c "UPDATE users set password=crypt('$WODAPIDBUSERPWD',gen_salt('bf')) where username='$WODAPIDBUSER';"
-	echo "Setup user $WODAPIDBADMIN"
-	psql --dbname=$POSTGRES_DB --username=postgres --host=localhost -c "UPDATE users set password=crypt('$WODAPIDBADMINPWD',gen_salt('bf')) where username='$WODAPIDBADMIN';"
-	echo "Setup user_roles table not done elsewhere"
-	psql --dbname=$POSTGRES_DB --username=postgres --host=localhost -c 'CREATE TABLE IF NOT EXISTS user_roles ("createdAt" timestamp DEFAULT current_timestamp, "updatedAt" timestamp DEFAULT current_timestamp, "roleId" integer CONSTRAINT no_null NOT NULL REFERENCES roles (id), "userId" integer CONSTRAINT no_null NOT NULL REFERENCES users (id));'
-	psql --dbname=$POSTGRES_DB --username=postgres --host=localhost -c 'ALTER TABLE user_roles ADD CONSTRAINT "ID_PKEY" PRIMARY KEY ("roleId","userId");'
-	# Get info on roles and users already declared
-	userroleid=`psql --dbname=$POSTGRES_DB --username=postgres --host=localhost -AXqtc "SELECT id FROM roles WHERE name='user';"`
-	moderatorroleid=`psql --dbname=$POSTGRES_DB --username=postgres --host=localhost -AXqtc "SELECT id FROM roles WHERE name='moderator';"`
-	adminroleid=`psql --dbname=$POSTGRES_DB --username=postgres --host=localhost -AXqtc "SELECT id FROM roles WHERE name='admin';"`
-	nbuser=`psql --dbname=$POSTGRES_DB --username=postgres --host=localhost -AXqtc "SELECT COUNT(id) FROM users;"`
-	moderatoruserid=`psql --dbname=$POSTGRES_DB --username=postgres --host=localhost -AXqtc "SELECT id FROM users WHERE username='$WODAPIDBUSER';"`
-	adminuserid=`psql --dbname=$POSTGRES_DB --username=postgres --host=localhost -AXqtc "SELECT id FROM users WHERE username='$WODAPIDBADMIN';"`
-	# Every user as a role of user so it's probably useless !
-	for (( i=$nbuser ; i>=1 ; i--)) do
-		psql --dbname=$POSTGRES_DB --username=postgres --host=localhost -c 'INSERT INTO user_roles ("roleId", "userId") VALUES ('$userroleid','$i');'
-	done
-	# Map the moderator user
-	psql --dbname=$POSTGRES_DB --username=postgres --host=localhost -c 'INSERT INTO user_roles ("roleId", "userId") VALUES ('$moderatorroleid','$moderatoruserid');'
-	# Map the admin user
-	psql --dbname=$POSTGRES_DB --username=postgres --host=localhost -c 'INSERT INTO user_roles ("roleId", "userId") VALUES ('$adminroleid','$adminuserid');'
-	echo "Starting API"
-	launch_with_pm2 $WODAPIDBDIR wod-$WODTYPE
+    echo "Launching docker PostgreSQL stack"
+    # Start the PostgreSQL DB stack
+	# We can use yq now as installed by ansible before
+    PGSQLDIR=`cat $WODAPIDBDIR/docker-compose.yml | yq '.services.db.environment.PGDATA' | sed 's/"//g'`
+    # We need to relog with sudo as $WODUSER so it's really in the docker group
+    # and be able to communicate with docker
+    # and we need to stop it before to be idempotent
+    # and we need to remove the data directory not done by the compose down
+    sudo su - $WODUSER -c "cd $WODAPIDBDIR ; docker compose down"
+    # That dir is owned by lxd, so needs root to remove
+    sudo su - -c "rm -rf $PGSQLDIR"
+    sudo su - $WODUSER -c "cd $WODAPIDBDIR ; docker compose config ; docker compose up -d"
+    POSTGRES_DB=`cat $WODAPIDBDIR/docker-compose.yml | yq '.services.db.environment.POSTGRES_DB' | sed 's/"//g'`
+    echo "Reset DB data"
+    npm run reset-data
+    echo "Setup user $WODAPIDBUSER"
+    psql --dbname=$POSTGRES_DB --username=postgres --host=localhost -c 'CREATE EXTENSION IF NOT EXISTS pgcrypto;'
+    psql --dbname=$POSTGRES_DB --username=postgres --host=localhost -c "UPDATE users set password=crypt('$WODAPIDBUSERPWD',gen_salt('bf')) where username='$WODAPIDBUSER';"
+    echo "Setup user $WODAPIDBADMIN"
+    psql --dbname=$POSTGRES_DB --username=postgres --host=localhost -c "UPDATE users set password=crypt('$WODAPIDBADMINPWD',gen_salt('bf')) where username='$WODAPIDBADMIN';"
+    echo "Setup user_roles table not done elsewhere"
+    psql --dbname=$POSTGRES_DB --username=postgres --host=localhost -c 'CREATE TABLE IF NOT EXISTS user_roles ("createdAt" timestamp DEFAULT current_timestamp, "updatedAt" timestamp DEFAULT current_timestamp, "roleId" integer CONSTRAINT no_null NOT NULL REFERENCES roles (id), "userId" integer CONSTRAINT no_null NOT NULL REFERENCES users (id));'
+    psql --dbname=$POSTGRES_DB --username=postgres --host=localhost -c 'ALTER TABLE user_roles ADD CONSTRAINT "ID_PKEY" PRIMARY KEY ("roleId","userId");'
+    # Get info on roles and users already declared
+    userroleid=`psql --dbname=$POSTGRES_DB --username=postgres --host=localhost -AXqtc "SELECT id FROM roles WHERE name='user';"`
+    moderatorroleid=`psql --dbname=$POSTGRES_DB --username=postgres --host=localhost -AXqtc "SELECT id FROM roles WHERE name='moderator';"`
+    adminroleid=`psql --dbname=$POSTGRES_DB --username=postgres --host=localhost -AXqtc "SELECT id FROM roles WHERE name='admin';"`
+    nbuser=`psql --dbname=$POSTGRES_DB --username=postgres --host=localhost -AXqtc "SELECT COUNT(id) FROM users;"`
+    moderatoruserid=`psql --dbname=$POSTGRES_DB --username=postgres --host=localhost -AXqtc "SELECT id FROM users WHERE username='$WODAPIDBUSER';"`
+    adminuserid=`psql --dbname=$POSTGRES_DB --username=postgres --host=localhost -AXqtc "SELECT id FROM users WHERE username='$WODAPIDBADMIN';"`
+    # Every user as a role of user so it's probably useless !
+    for (( i=$nbuser ; i>=1 ; i--)) do
+        psql --dbname=$POSTGRES_DB --username=postgres --host=localhost -c 'INSERT INTO user_roles ("roleId", "userId") VALUES ('$userroleid','$i');'
+    done
+    # Map the moderator user
+    psql --dbname=$POSTGRES_DB --username=postgres --host=localhost -c 'INSERT INTO user_roles ("roleId", "userId") VALUES ('$moderatorroleid','$moderatoruserid');'
+    # Map the admin user
+    psql --dbname=$POSTGRES_DB --username=postgres --host=localhost -c 'INSERT INTO user_roles ("roleId", "userId") VALUES ('$adminroleid','$adminuserid');'
+    echo "Starting API"
+    launch_with_pm2 $WODAPIDBDIR wod-$WODTYPE
 elif [ $WODTYPE = "frontend" ]; then
-	cd $WODFEDIR
-	cat > .env << EOF
+    cd $WODFEDIR
+    cat > .env << EOF
 GATSBY_WORKSHOP_API_ENDPOINT=http://$WODAPIDBFQDN:$WODAPIDBPORT
 GATSBY_USERNAME=''
 GATSBY_PASSWORD=''
@@ -343,17 +356,17 @@ GATSBY_NEWSLETTER_API=''
 WODUID=`id -u`
 WODGID=`id -g`
 EOF
-	echo "Launching npm install..."
-	npm install
-	echo "Patching package.json to allow listening on the right host:port"
-	perl -pi -e "s|gatsby develop|gatsby develop -H $WODFEFQDN -p $WODFEPORT|" package.json
-	echo "Start the Frontend server"
-	launch_with_pm2 $WODFEDIR wod-$WODTYPE
+    echo "Launching npm install..."
+    npm install
+    echo "Patching package.json to allow listening on the right host:port"
+    perl -pi -e "s|gatsby develop|gatsby develop -H $WODFEFQDN -p $WODFEPORT|" package.json
+    echo "Start the Frontend server"
+    launch_with_pm2 $WODFEDIR wod-$WODTYPE
 fi
 
 if [ $WODTYPE != "appliance" ]; then
-	cd $SCRIPTDIR/../ansible
+    cd $SCRIPTDIR/../ansible
 
-	ansible-playbook -i inventory $WODPRIVINV --limit $PBKDIR $ANSPLAYOPT $ANSPRIVOPT check_$WODTYPE.yml
+    ansible-playbook -i inventory $WODPRIVINV --limit $PBKDIR $ANSPLAYOPT $ANSPRIVOPT check_$WODTYPE.yml
 fi
 date
